rename folder flywheel --> flywheels
