+++ src/main/java/frc/robot/RobotContainer.java
@@ -42,10 +42,10 @@ import frc.robot.subsystems.shooter.BeamBreakIO;
 import frc.robot.subsystems.shooter.BeamBreakIORev;
 import frc.robot.subsystems.shooter.BeamBreakIOSim;
 import frc.robot.subsystems.shooter.Shooter;
-import frc.robot.subsystems.shooter.flywheel.FlywheelsIOSim;
-import frc.robot.subsystems.shooter.flywheel.FlywheelsIOSparkMax;
-import frc.robot.subsystems.shooter.flywheel.Flywheels;
-import frc.robot.subsystems.shooter.flywheel.FlywheelsIO;
+import frc.robot.subsystems.shooter.flywheels.Flywheels;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIO;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIOSim;
+import frc.robot.subsystems.shooter.flywheels.FlywheelsIOSparkMax;
 import frc.robot.subsystems.shooter.pivot.Pivot;
 import frc.robot.subsystems.shooter.pivot.PivotIO;
 import frc.robot.subsystems.shooter.pivot.PivotIOSim;
