+++ src/main/java/frc/robot/subsystems/shooter/flywheels/FlywheelsIO.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.shooter.flywheel;
+package frc.robot.subsystems.shooter.flywheels;
 
 import org.littletonrobotics.junction.AutoLog;
